The method for add a value on the boar was added
This method add a value on the board in the cell number that the player specifies.
As the number of cell is similar to a label, the add_value method call tre methods:
find_x_coordinate, find_y_coordinate and empty_cell?, the first two find the
coordinates (y,x) in the board that correspond to the cell number chosen by the user.
And the last one verifies that the cell is empty.