The put_tap, play_again, ask_board_size classes was added
The first one ask to the player in turn for a position in the
board and call another method that evaluates if the position
exists or is available. The second method ask to the player if he wants to start a new game. And the last one ask to the player the size
of the board and evaluates that the size is greather or equal to 3.