The Tic Tac Toe game is functional
In this commit was added more methods to the main class.
Among these was added the start_game, this is in charge of
simulates a Tic Tac Toe game.
Also was added the winner_is method, this puts in the screen the
name of the winner when there is a winner.
Methods that display the board in the screen after each player turn
were added. The change_turns_players take the control of players turn.